Huge refactoring of window management and backend
First of all, I have created a class for the db management so MainWindow can concentrate on the GUI instead of
database stuff.

Second of all, window management is a lot nicer now. No more 'Open' dialog on application startup, since
windows can have proper 'No file opened yet' state. When you open a file in this state, the window will
open the file. If you open outside this state, a new window will be opened for the other file.